requirements.txt
pandas==1.3.3
scikit_learn==1.0.2
streamlit==1.4.0
google-cloud-bigquery==2.32.0
